Create 13)Custom Search Function in Python
This Python script demonstrates the implementation of a custom search function that searches for an element within a list. It offers two variations of the search function, one that returns a boolean value indicating whether the element is found and another that provides the position (index) of the element if found. These functions can be used to search for specific values within a list and are helpful in various programming scenarios.